fix: rename PgMWSTConnection file
