+++ acapy_wallet_upgrade/strategies.py
@@ -14,7 +14,7 @@ import msgpack
 import nacl.pwhash
 
 from acapy_wallet_upgrade.error import UpgradeError
-from acapy_wallet_upgrade.pg_connection_mwst_profiles import PgMWSTConnection
+from acapy_wallet_upgrade.pg_mwst_connection import PgMWSTConnection
 
 from .db_connection import DbConnection, Wallet
 from .pg_connection import PgConnection, PgWallet
